Rename proto organization to trace_machina
