+++ cas/worker/tests/utils/mock_running_actions_manager.rs
@@ -21,7 +21,7 @@ use tokio::sync::mpsc;
 use action_messages::ActionResult;
 use common::DigestInfo;
 use error::{make_input_err, Error};
-use proto::com::github::allada::turbo_cache::remote_execution::StartExecute;
+use proto::com::github::trace_machina::turbo_cache::remote_execution::StartExecute;
 use running_actions_manager::{Metrics, RunningAction, RunningActionsManager};
 
 #[derive(Debug)]
