+++ proto/com/github/trace_machina/turbo_cache/remote_execution/worker_api.proto
@@ -14,7 +14,7 @@
 
 syntax = "proto3";
 
-package com.github.allada.turbo_cache.remote_execution;
+package com.github.trace_machina.turbo_cache.remote_execution;
 
 import "build/bazel/remote/execution/v2/remote_execution.proto";
 import "google/protobuf/empty.proto";
@@ -85,7 +85,7 @@ message SupportedProperties {
     /// the unsupported properties.
     ///
     /// The details on how to use this property can be found here:
-    /// https://github.com/allada/turbo-cache/blob/c91f61edf182f2b64451fd48a5e63fa506a43aae/config/cas_server.rs
+    /// https://github.com/trace_machina/turbo-cache/blob/c91f61edf182f2b64451fd48a5e63fa506a43aae/config/cas_server.rs
     repeated build.bazel.remote.execution.v2.Platform.Property properties = 1;
     reserved 2; // NextId.
 }
